+++ src/main/java/bta/aether/AetherBlockTags.java
@@ -1,0 +1,19 @@
+package bta.aether;
+
+import net.minecraft.core.block.Block;
+import net.minecraft.core.data.tag.Tag;
+
+public class AetherBlockTags {
+
+    public static Tag<Block> MINEABLE_BY_AETHER_PICKAXE = Tag.of("mineable_by_aether_pickaxe");
+
+    public static Tag<Block> MINEABLE_BY_AETHER_AXE = Tag.of("mineable_by_aether_axe");
+
+    public static Tag<Block> MINEABLE_BY_AETHER_SHOVEL = Tag.of("mineable_by_aether_shovel");
+
+    public static Tag<Block> MINEABLE_BY_AETHER_SWORD = Tag.of("mineable_by_aether_sword");
+
+    public static Tag<Block> GROWS_AETHER_FLOWERS = Tag.of("grows_flowers");
+
+
+}
