+++ src/main/java/bta/aether/mixin/ItemBucketMixin.java
@@ -1,10 +1,11 @@
 package bta.aether.mixin;
 
-import bta.aether.Aether;
 import bta.aether.block.AetherBlocks;
 import net.minecraft.core.HitResult;
 import net.minecraft.core.block.Block;
+import net.minecraft.core.block.tag.BlockTags;
 import net.minecraft.core.entity.player.EntityPlayer;
+import net.minecraft.core.item.Item;
 import net.minecraft.core.item.ItemBucket;
 import net.minecraft.core.item.ItemStack;
 import net.minecraft.core.sound.SoundType;
@@ -12,55 +13,69 @@ import net.minecraft.core.util.helper.MathHelper;
 import net.minecraft.core.util.phys.Vec3d;
 import net.minecraft.core.world.World;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(value = ItemBucket.class, remap = false)
-public class ItemBucketMixin {
+public class ItemBucketMixin extends Item{
 
-    private final int idToPlace;
-    public ItemBucketMixin(int idToPlace) {
-        this.idToPlace = idToPlace;
+    @Shadow
+    private int idToPlace;
+
+    public ItemBucketMixin(String name, int id) {
+        super(name, id);
     }
 
-    @Inject(
-            method = "onItemRightClick(Lnet/minecraft/core/item/ItemStack;Lnet/minecraft/core/world/World;Lnet/minecraft/core/entity/player/EntityPlayer;)Lnet/minecraft/core/item/ItemStack;",
-            at = @At(value = "TAIL")
-    )
-    public void aetherAerogel(ItemStack stack, World world, EntityPlayer player, CallbackInfoReturnable<ItemStack> cir) {
-        float f = 1.0F;
-        float f1 = player.xRotO + (player.xRot - player.xRotO) * f;
-        float f2 = player.yRotO + (player.yRot - player.yRotO) * f;
-        double posX = player.xo + (player.x - player.xo) * (double)f;
-        double posY = player.yo + (player.y - player.yo) * (double)f + 1.62 - (double)player.heightOffset;
-        double posZ = player.zo + (player.z - player.zo) * (double)f;
-        float f3 = MathHelper.cos(-f2 * 0.01745329F - 3.141593F);
-        float f4 = MathHelper.sin(-f2 * 0.01745329F - 3.141593F);
-        float f5 = -MathHelper.cos(-f1 * 0.01745329F);
-        float f6 = MathHelper.sin(-f1 * 0.01745329F);
-        float f7 = f4 * f5;
-        float f9 = f3 * f5;
-        double d3 = 5.0;
-        Vec3d vec3d = Vec3d.createVector(posX, posY, posZ);
-        Vec3d vec3d1 = vec3d.addVector((double)f7 * d3, (double)f6 * d3, (double)f9 * d3);
-        HitResult rayTraceResult = world.checkBlockCollisionBetweenPoints(vec3d, vec3d1, this.idToPlace == 0);
-        if (rayTraceResult != null && rayTraceResult.hitType == HitResult.HitType.TILE) {
-            int x = rayTraceResult.x;
-            int y = rayTraceResult.y;
-            int z = rayTraceResult.z;
-            if (y >= 0 && y < world.getHeightBlocks()) {
-                if (world.isAirBlock(x, y, z) || !world.getBlockMaterial(x, y, z).isSolid()) {
-                    if (world.dimension == Aether.dimensionAether && this.idToPlace == Block.fluidLavaFlowing.id) {
-                        world.playSoundEffect(SoundType.WORLD_SOUNDS, (double) z + 0.5, (double) y + 0.5, (double) x + 0.5, "random.fizz", 0.5F, 2.6F + (world.rand.nextFloat() - world.rand.nextFloat()) * 0.8F);
+    @Inject(method = "onItemRightClick", at = @At("HEAD"), cancellable = true)
+    public void callOnItemRightClick(ItemStack stack, World world, EntityPlayer player, CallbackInfoReturnable<ItemStack> info) {
+        if (world.dimension.languageKey.equals("aether") && Block.getBlock(idToPlace).hasTag(BlockTags.IS_LAVA)) {
 
-                        for (int l = 0; l < 8; ++l) {
-                            world.spawnParticle("largesmoke", (double) x + Math.random(), (double) y + Math.random(), (double) z + Math.random(), 0.0, 0.0, 0.0);
-                                world.setBlockWithNotify(x, y, z, AetherBlocks.aerogel.id);
-                        }
-                    }
+            float pitch = player.xRotO + (player.xRot - player.xRotO);
+            float yaw = player.yRotO + (player.yRot - player.yRotO);
+
+            double playerPosX = player.xo + (player.x - player.xo);
+            double playerPosY = player.yo + (player.y - player.yo) + 1.62 - (double)player.heightOffset;
+            double playerPosZ = player.zo + (player.z - player.zo);
+
+            Vec3d playerPosition = Vec3d.createVector(playerPosX, playerPosY, playerPosZ);
+
+            float pitchCos = MathHelper.cos(-yaw * 0.01745329F - 3.141593F);
+            float pitchSin = MathHelper.sin(-yaw * 0.01745329F - 3.141593F);
+            float yawCos = -MathHelper.cos(-pitch * 0.01745329F);
+            float yawSin = MathHelper.sin(-pitch * 0.01745329F);
+
+            float pitchYawCos = pitchSin * yawCos;
+            float pitchYawSin = pitchCos * yawCos;
+
+            double distance = 7.0f;
+            Vec3d endPosition = playerPosition.addVector((double)pitchYawCos * distance, (double)yawSin * distance, (double)pitchYawSin * distance);
+
+            HitResult rayTraceResult = world.checkBlockCollisionBetweenPoints(playerPosition, endPosition, this.idToPlace == 0);
+
+            if (rayTraceResult != null && rayTraceResult.hitType == HitResult.HitType.TILE) {
+
+                int x = rayTraceResult.side.getOffsetX() + rayTraceResult.x;
+                int y = rayTraceResult.side.getOffsetY() + rayTraceResult.y;
+                int z = rayTraceResult.side.getOffsetZ() + rayTraceResult.z;
+
+                player.swingItem();
+                world.playSoundEffect(SoundType.WORLD_SOUNDS, (double) x + 0.5, (double) y + 0.5, (double) z + 0.5, "fire.ignite", 1.0F, itemRand.nextFloat() * 0.4F + 0.8F);
+                for (int l = 0; l < 8; ++l) {
+                    double angle = Math.toRadians(l * 45);
+                    world.spawnParticle("largesmoke", (double) x + 0.5, (double) y, (double) z + 0.5, -(Math.cos(angle) * 2) / 20.0, 0.03, -(Math.sin(angle) * 2)/ 20.0);
                 }
+
+                world.setBlockWithNotify(x, y, z, AetherBlocks.aerogel.id);
+
+                if (player.getGamemode().consumeBlocks()) {
+                    info.setReturnValue(new ItemStack(Item.bucket));
+                    return;
+                }
+                info.setReturnValue(stack);
             }
         }
     }
+
 }
